Put usecases into __init__
