+++ src/features/__init__.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from src.features.ingestion.usecases import Ingestion
+from src.features.ingestion import Ingestion
 
 __all__ = [
     "Features",
